refactor
- changed LDL_DEBUG, LDL_INFO, LDL_ERROR to not insert __FUNCTION__ into
format string
- changed LDL_DEBUG, LDL_INFO, LDL_ERROR to not take APP argument
- reworked log messages and levels
- added LDL_ENABLE_FAST_DEBUG build option which removes extra delays
and dithering required
  for use in the field but which slow down development
- changed band timers to use gps timebase (1/256Hz timebase)
- changed time tracker to use gps timebase
- removed logic for detecting and handling late RX windows since this
won't normally occur
- removed LDL_MAC_timeSinceValidDownlink()
- removed unused ldl_mac fields
- removed events that are not useful to the application (and replaced
with log messages)
    - removed LDL_MAC_CHIP_ERROR event
    - removed LDL_MAC_RESET event
    - removed LDL_MAC_RX1_SLOT event
    - removed LDL_MAC_RX2_SLOT event
    - removed LDL_MAC_DOWNSTREAM
    - removed LDL_MAC_TX_BEGIN
    - removed LDL_MAC_TX_COMPLETE
- added examples/avr to keep an eye on whether LDL still fits on a 328P
- changed LDL_Radio_getAirTime() to return result in milliseconds and
not in a timebase of your choice
- fixed most required MISRA 2012 warnings

+++ include/ldl_radio_debug.h
@@ -51,7 +51,7 @@ struct ldl_radio_debug_log {
 };
 
 void LDL_Radio_debugLogReset(struct ldl_radio *self);
-void LDL_Radio_debugLogPush(struct ldl_radio *self, uint8_t opcode, const uint8_t *data, uint8_t size);
+void LDL_Radio_debugLogPush(struct ldl_radio *self, uint8_t opcode, const uint8_t *data, size_t size);
 void LDL_Radio_debugLogFlush(struct ldl_radio *self, const char *fn);
 
 #ifdef __cplusplus
